+++ go.mod
@@ -1,8 +1,8 @@
-module github.com/containeroo/certinfo
+module certinfo
 
 go 1.16
 
 require (
 	github.com/carlmjohnson/errutil v0.21.2
-	github.com/carlmjohnson/flagext v0.21.0
+	github.com/spf13/cobra v1.2.1
 )
